requirements.txt
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask-Dance==6.0.0
Flask-Login==0.6.1
Flask-OAuth==0.12
Flask-Pydantic==0.9.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauth2==1.9.0.post1
oauthlib==3.2.0
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
SQLAlchemy==1.4.39
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.10
URLObject==2.4.3
Werkzeug==2.1.2
zipp==3.8.1
